Improved Zotero Caching
We use the 'version' of the Zotero Collection Spec to manage the caching behavior of items throughout the system. We had been using exclusively the `version` column of items to provide this value, but this had a couple of issues:

- the version column is set by syncing to the server, which means that it will not be set if there is no sync, but even if there is sync, the version isn't set until the server has been updating (often many seconds after it has been written).

- the version column of items doesn't capture all the ways that the cache should be broken, especially changes to collections or collection ownership of items.

This change does two major things:

1) Switch to using the clientDateModified field to provide the version. To support this, we need to switch the version of the client spec (and all intermediaries) to doubles.

2) Check both the collection clientDateModified and the item clientDateModified dates to determine the version.

+++ src/cpp/session/modules/zotero/ZoteroCollections.cpp
@@ -67,7 +67,7 @@ FilePath collectionsCacheDir(const std::string& type, const std::string& context
 struct IndexedCollection
 {
    bool empty() const { return file.empty(); }
-   int version;
+   double version;
    std::string file;
    std::string key;
    std::string parentKey;
@@ -92,7 +92,7 @@ std::map<std::string,IndexedCollection> collectionsCacheIndex(const FilePath& ca
 
                json::Object entryJson = member.getValue().getObject();
                IndexedCollection coll;
-               coll.version = entryJson[kVersion].getInt();
+               coll.version = entryJson[kVersion].getDouble();
                coll.file = entryJson[kFile].getString();
                coll.key = entryJson[kKey].getString();
                coll.parentKey = entryJson[kParentKey].getString();
@@ -142,7 +142,7 @@ Error readCollection(const FilePath& filePath, ZoteroCollection* pCollection)
       return error;
 
    pCollection->name = collectionJson[kName].getString();
-   pCollection->version = collectionJson[kVersion].getInt();
+   pCollection->version = collectionJson[kVersion].getDouble();
    pCollection->key = collectionJson[kKey].getString();
    pCollection->parentKey = collectionJson[kParentKey].getString();
    pCollection->items = collectionJson[kItems].getArray();
@@ -326,7 +326,7 @@ const char * const kKey = "key";
 const char * const kParentKey = "parentKey";
 const char * const kItems = "items";
 
-const int kNoVersion = -1;
+const double kNoVersion = -1;
 
 ZoteroCollectionSpec findParentSpec(const ZoteroCollectionSpec& spec, const ZoteroCollectionSpecs& specs)
 {
